FIX: improves resilience if topic.category would end up being nil
This is not supposed to happen with the normak behavior, however it can happen in tests for example, and other plugin's tests could fail when tested alongside zendesk plugin.

This commit essentially ensures it works even if `saved_changes[:category_id]` value was not two ids, eg: `[2, nil]`

The commit also adds a spec to this untested part of the plugin.