UX: Improve setting names and descriptions
The zendesk_enable_all_categories and zendesk_enabled_categories setting names and descriptions were not as clear as the could be.

For context, adjusting these settings only affects the autogeneration of tickets. When they are not enabled/populated, there is still a button below the topic footer to manually create tickets.